fea/view profile by id, roll back deleted order, view deleted order

+++ KoiDeliveryOrderingSystem-Backend/src/main/java/com/swp391/group7/KoiDeliveryOrderingSystem/service/PackageService.java
@@ -119,10 +119,13 @@ public class PackageService {
     public PackageResponse deletePackage(Integer id) {
         Users users = accountUtils.getCurrentUser();
         if (users == null) {
-            throw new AppException(ErrorCode.USER_NOT_EXISTED);
+            throw new AppException(ErrorCode.NOT_LOGIN);
         }
         Package packages = packageRepository.findById(id)
                 .orElseThrow(() -> new AppException(ErrorCode.PACKAGE_NOT_FOUND));
+        if (packages.getPackageStatus().equals(PackageStatusEnum.PACKED)) {
+            throw new AppException(ErrorCode.NOT_DELETE_PACKED);
+        }
         packageRepository.delete(packages);
         return convertToPackageResponse(packages);
     }
